Add Node 16 to test matrix and remove Node 10 (#1318)
Removes Node 10 from the test matrix since it is no longer supported and adds Node 16. This changes only apply to v6 since mgmt. plane packages already moved or are moving out from v4.

It also removes the folder eng\pipelines\ and all its contents since this is a duplicate of scripts\ and is no longer in used.