[autorest.typescript] Allow generating ESM-style imports (#2925)
This PR adds the `module-kind` flag to autorest options, allowing the generated code to have ESM-style imports/exports. Those may be generated with `.js` extension and/or `index.js` for directory-style imports.

The default remains CJS in autorest, but the idea is to then add this option to each HLC package that migrates to ESM in our repo.

The yaml would look like:
```yaml
module-kind: esm
```

For any package that has already migrated